requirements.txt
about-time==3.1.1
alive-progress==2.4.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==38.0.4
dataclasses-json==0.5.7
grapheme==0.6.0
h11==0.14.0
idna==3.3
instaloader==4.9.5
marshmallow==3.19.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.20.0
python-facebook-api==0.15.0
requests==2.27.1
requests-oauthlib==1.3.1
selenium==4.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing-inspect==0.8.0
typing_extensions==4.4.0
urllib3==1.26.9
webdriver-manager==3.8.5
wsproto==1.2.0
